# backend/docs/swagger.yaml
basePath: /
definitions:
...
  entity.Organization:
    properties:
      created_at:
        type: string
      created_by:
        description: user_id
        type: string
      description:
        type: string
      display_name:
        description: 显示名称
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        description: 唯一标识名称
        type: string
      settings:
        additionalProperties: true
        description: 组织配置
        type: object
      updated_at:
        type: string
    type: object
...
  entity.Project:
    properties:
      created_at:
        type: string
      created_by:
        description: user_id
        type: string
      description:
        type: string
      display_name:
        description: 显示名称
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_default:
        description: 是否为默认项目
        type: boolean
      name:
        description: 项目名称
        type: string
      org_id:
        description: 所属组织ID
        type: integer
      organization:
        allOf:
        - $ref: '#/definitions/entity.Organization'
        description: 关联
      settings:
        additionalProperties: true
        description: 项目配置
        type: object
      updated_at:
        type: string
    type: object
...
  handlers.CreateOrganizationRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      name:
        type: string
      settings:
        additionalProperties: true
        type: object
    required:
    - name
    type: object
  handlers.CreateProjectRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      name:
        type: string
      org_id:
        type: integer
      settings:
        additionalProperties: true
        type: object
    required:
    - name
    - org_id
    type: object
...
  handlers.UpdateOrganizationRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      is_active:
        type: boolean
      settings:
        additionalProperties: true
        type: object
    type: object
  handlers.UpdateProjectRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      is_active:
        type: boolean
      settings:
        additionalProperties: true
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@iac-platform.com
    name: API Support
  description: Infrastructure as Code Platform API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: IAC Platform API
  version: "1.0"
paths:
...
  /api/v1/organizations:
    get:
      parameters:
      - description: 是否启用
        in: query
        name: is_active
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Organization'
            type: array
      summary: 列出所有组织
      tags:
      - Organization
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建组织请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Organization'
      summary: 创建组织
      tags:
      - Organization
  /api/v1/organizations/{id}:
    delete:
      parameters:
      - description: 组织ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除组织
      tags:
      - Organization
    get:
      parameters:
      - description: 组织ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Organization'
      summary: 获取组织详情
      tags:
      - Organization
    put:
      parameters:
      - description: 组织ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新组织请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateOrganizationRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新组织
      tags:
      - Organization
...
  /api/v1/projects:
    get:
      parameters:
      - description: 组织ID
        in: query
        name: org_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Project'
            type: array
      summary: 列出组织的所有项目
      tags:
      - Project
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建项目请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Project'
      summary: 创建项目
      tags:
      - Project
  /api/v1/projects/{id}:
    delete:
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除项目
      tags:
      - Project
    get:
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Project'
      summary: 获取项目详情
      tags:
      - Project
    put:
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新项目请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProjectRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新项目
      tags:
      - Project
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"